Move components into correct directory
